# Conflicts:
#	database/schemas/com.ut.database.database.CloudLockRoomDatabase/3.json
#	database/schemas/com.ut.database.database.CloudLockRoomDatabase/4.json
#	database/src/main/java/com/ut/database/database/CloudLockRoomDatabase.java
#	lib_networkapi/src/main/java/com/example/operation/MyRetrofit.java
#	module_lock/src/main/java/com/ut/module_lock/fragment/LockListFragment.java

+++ database/src/main/java/com/ut/database/dao/LockKeyDao.java
@@ -29,7 +29,7 @@ public interface LockKeyDao {
     void insertAll(LockKey... lockKeys);
 
     @Query("SELECT * FROM lock_key WHERE groupId = :groupId ORDER BY userType ASC")
-    LiveData<List<LockKey>> getLockByGroupId(int groupId);
+    LiveData<List<LockKey>> getLockByGroupId(long groupId);
 
     @Query("SELECT * FROM lock_key WHERE name LIKE :name ORDER BY userType ASC")
     LiveData<List<LockKey>> getLockByName(String name);
